+++ shell.nix
@@ -2,9 +2,9 @@ with (import (builtins.fetchTarball {
   # Descriptive name to make the store path easier to identify
   name = "typeit-python38";
   # Commit hash for nixos-unstable as of 2019-10-27
-  url = https://github.com/NixOS/nixpkgs/archive/4aaf2ad527e123e99ecb993130e8458125a6a42f.tar.gz;
+  url = https://github.com/NixOS/nixpkgs-channels/archive/f601ab37c2fb7e5f65989a92df383bcd6942567a.tar.gz;
   # Hash obtained using `nix-prefetch-url --unpack <url>`
-  sha256 = "0h9lbijvzbm30dlzm7y4v6iah5d8bxbw8i6ipsfy6fapyf2l79z0";
+  sha256 = "0ikhcmcc29iiaqjv5r91ncgxny2z67bjzkppd3wr1yx44sv7v69s";
 }) {});
 
 # Make a new "derivation" that represents our shell
@@ -35,9 +35,18 @@ stdenv.mkDerivation {
         # set SOURCE_DATE_EPOCH so that we can use python wheels
         export SOURCE_DATE_EPOCH=$(date +%s)
 
-        virtualenv $PWD/.venv
-        export PATH=$PWD/.venv/bin:$PATH
+        VENV_DIR=$PWD/.venv
 
+        export PATH=$VENV_DIR/bin:$PATH
+        export PYTHONPATH=""
         export LANG=en_US.UTF-8
+
+        # https://python-poetry.org/docs/configuration/
+        export PIP_CACHE_DIR=$PWD/.local/pip-cache
+
+        # Setup virtualenv
+        if [ ! -d $VENV_DIR ]; then
+            virtualenv $PWD/.venv
+        fi
     '';
 }