+++ tests/test_cli.py
@@ -5,7 +5,7 @@ from unittest import mock
 from typeit.cli import main
 
 
-TYPE_REF = 'mk_main, serialize_main = type_constructor ^ Main'
+TYPE_REF = 'mk_main, serialize_main = TypeConstructor ^ Main'
 
 @pytest.mark.parametrize("stdin_data, reference_snippet", [
     ("""{"x": "Hello", "y": "World", "z": {"a": 1}}""", TYPE_REF),
@@ -15,15 +15,15 @@ TYPE_REF = 'mk_main, serialize_main = type_constructor ^ Main'
     z:
       a: 1
     """, TYPE_REF),
-    ("""1""", "mk_int, serialize_int = type_constructor ^ int"),
-    ("""1.0""", "mk_float, serialize_float = type_constructor ^ float"),
-    (''' "1" ''', "mk_str, serialize_str = type_constructor ^ str"),
-    (''' true ''', "mk_bool, serialize_bool = type_constructor ^ bool"),
-    (''' null ''', "mk_none, serialize_none = type_constructor ^ None"),
-    (''' [] ''', "Main = Sequence[Any]\n\n\nmk_main, serialize_main = type_constructor ^ Main"),
-    (''' [[]] ''', "Main = Sequence[Sequence[Any]]\n\n\nmk_main, serialize_main = type_constructor ^ Main"),
-    (''' [[1]] ''', "Main = Sequence[Sequence[int]]\n\n\nmk_main, serialize_main = type_constructor ^ Main"),
-    (''' [[ {"x": {"y": null}} ]] ''', "class Main(NamedTuple):\n    x: X\n\n\nmk_main, serialize_main = type_constructor ^ Main"),
+    ("""1""", "mk_int, serialize_int = TypeConstructor ^ int"),
+    ("""1.0""", "mk_float, serialize_float = TypeConstructor ^ float"),
+    (''' "1" ''', "mk_str, serialize_str = TypeConstructor ^ str"),
+    (''' true ''', "mk_bool, serialize_bool = TypeConstructor ^ bool"),
+    (''' null ''', "mk_none, serialize_none = TypeConstructor ^ None"),
+    (''' [] ''', "Main = Sequence[Any]\n\n\nmk_main, serialize_main = TypeConstructor ^ Main"),
+    (''' [[]] ''', "Main = Sequence[Sequence[Any]]\n\n\nmk_main, serialize_main = TypeConstructor ^ Main"),
+    (''' [[1]] ''', "Main = Sequence[Sequence[int]]\n\n\nmk_main, serialize_main = TypeConstructor ^ Main"),
+    (''' [[ {"x": {"y": null}} ]] ''', "class Main(NamedTuple):\n    x: X\n\n\nmk_main, serialize_main = TypeConstructor ^ Main"),
 ])
 def test_cli(stdin_data, reference_snippet):
     out_channel = io.StringIO()