+++ typeit/__init__.py
@@ -1,6 +1,6 @@
 from . import flags
 from . import sums
-from .combinator.constructor import type_constructor
+from .combinator.constructor import type_constructor, TypeConstructor
 from .schema.errors import Error
 
-__all__ = ('type_constructor', 'flags', 'sums', 'Error')
+__all__ = ('TypeConstructor', 'type_constructor', 'flags', 'sums', 'Error')